requirements.txt
asgiref==3.9.1
av==15.1.0
cffi==2.0.0
cryptography==45.0.7
decorator==5.2.1
Django==5.2.6
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
imageio==2.37.0
imageio-ffmpeg==0.6.0
moviepy==2.2.1
numpy==2.3.2
pillow==11.3.0
proglog==0.1.12
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
tqdm==4.67.1
